# DOC: Document performance cliff for scipy.sparse.random