<!-- 
Thanks for contributing a pull request! Please ensure that
your PR satisfies the checklist before submitting:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#checklist-before-submitting-a-pr

Also, please name and describe your PR as you would write a
commit message:
http://scipy.github.io/devdocs/dev/contributor/development_workflow.html#writing-the-commit-message

Depending on your changes, you can skip CI operations and save time and energy: 
http://scipy.github.io/devdocs/dev/contributor/continuous_integration.html#skipping

Note that we are a team of volunteers; we appreciate your
patience during the review process.

Again, thanks for contributing!
-->

#### Reference issue
<!--Example: Closes gh-WXYZ.-->

Documents current state of https://github.com/scipy/scipy/issues/9699

#### What does this implement/fix?

This documents that the default arguments for `sparse.random` perform poorly, and how to get much better performance.

#### Additional information

Example of performance:

```python
In [7]: %time result = sparse.random(100_000, 10000, density=0.01, random_state=np.random.default_rng())
CPU times: user 1.05 s, sys: 143 ms, total: 1.2 s
Wall time: 1.21 s

In [10]: %time result = sparse.random(100_000, 10000, density=0.01)
CPU times: user 1min 25s, sys: 2.43 s, total: 1min 27s
Wall time: 1min 27s
```